requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.7.2
opencv-python-headless==4.10.0.84
easyocr==1.7.1
sympy==1.13.1
Pillow==10.4.0
numpy==1.26.4
reportlab==4.2.2
python-docx==1.1.2
requests==2.32.3
gunicorn==22.0.0
